Changed handling of Operator/Jordan-Wigner-strings
The function `CouplingModel.add_site` has now an option (activated by
default!) to extend the Jordan-Wigner string to the "left" of the two
operators.
This is necessary to correctly handle Jordan-Wigner strings on
different geometries (e.g. a snake in 2D)!

Default-value for `str_on_first` in MPS.correlation_function changed!