!!! changed meaning of truncation parameter `trunc_err`
before this commit: dicard singular values as long as
    sum_{i discarded} S[i]**2 < trunc_err
after this commit: dicard singular values as long as
    sum_{i discarded} S[i]**2 < trunc_err**2
Reasoning:
No one ever looks up these precise definitions.
Instead, one reads/remembers "`svd_min` is the cutoff for singular values",
and then chooses `trunc_cut` on the same order of magnitude.

If the singular values decay quickly, the sum is dominated by the
maximal S[i] which is discarded.
In the old definition, this discards singular values (almost) as big as sqrt(trunc_err);
with the new definition its sqrt(trunc_err**2) == trunc_err.
This means one can choose them on the same order of magnitude, namely
the desired precision of the final result...