BACKWARDS INCOMPATIBLE change in MPS.from_product_state and MPS.from_Bflat
Sorting a physical LegCharge might require a permutation of the basis in
a local Site. If e.g. ``p_state=[i]*L`` was given to the MPS, the
resulting MPS was nonzero in B[i, 0, 0], but the meaning of
this changed when a nonzero permutation was necessary in the Site (i.e.
when the operators where permuted), e.g. the expectation value of the
operators might suddenly change.
One way to avoid this is to use strings labeling the states.
(This was documented, but somehow not implemented. I implemented it...)
Another way is to apply the same permutation used for the
operators to the state (when generating the site) also on the MPS.
This second way is what the new permutation=True argument
enables now by default in both MPS.from_product_state and MPS.from_Bflat.