BACKWARDS INCOMPATIBLE: generalize DoubleSite to GroupedSite
Generalize and rename :class:`tenpy.networks.site.DoubleSite` to :class:`tenpy.networks.site.GroupedSite`,
to allow for an arbitrary number of sites to be grouped.
Arguments ``site0, site1, label0, label1`` of the __init__ can be replaced with ``[site0, site1], [label0, label1]``
and ``op0, op1`` of the `kronecker_product` with ``[op0, op1]``; this will recover the functionality of the `DoubleSite`.
Within TeNPy, the DoubleSite was so far only used in tenpy.models.spins_nnn.