found the cases in which we have cython memory leaks
The current cython produces memory leaks when using the
`cdef ndarray[intp_t, ndim=2] a = np.zeros([2, 0])`
syntax with an array `a` which has shape 0 in the second dimension!
In TeNPy this happened often when we have no charges, ``qnumber=0``!
Introduced a few more checks for this, which should also speed up things
in this case.