testing: use pytest instead of nose
Nose is not developed actively anymore, so we switch to pytest.
Pytest has a better introspection method giving nicer error messages,
and is able to run the nose tests (up to tests with "yield", which now need use `parametrize`).
Moreover, we will be able to integrate the tests into the setup.py