Issue #94 for npc.inner
npc.inner takes the axes in the same order for both arrays and contracts them.
This is bad, if the arrays coincide, except that one is transposed - and we only
see it by the labels, but the legs actually fit.
This commit introduces the options ``axes='range'`` with the old behaviour,
and the new option ``axes='labels'``, which indicates that the arrays
have the same labels and should be transposed to fit together accordingly.

Closes #94.