FIX: allow to return truncation errors down to 1.e-30
A sublte precision error:
Before, the TruncationError.eps, which is supposed to be
`sum(discarded singular values squared)`, was returned as 0 if
the exact value was below machine precision. The reason was, that
it was evaluated as `1. - sum(kept singular values squared)`,
and `1. - (1. - 1.e-20)` evaluates to zero.

Related to issue #95, since this leads to the fact that `P_tol` is never
updated with `P_tol_to_trunc` in DMRG.