don't raise an Error if numpy version changed since installation.
What we need is ABI compatibility. And as far as I understand, this
should be given for all increasing minor releases,
e.g. compiling with numpy v1.11 will allow to use numpy 1.13, but not numpy2.0.
Anyways, it seeems checking the numpy version is too
restrictive - as long as everything works, it works, right?