TimeEvolutionAlgorithm and Algorithm.get_resume_data()
There should be a single dictionary containing all the data necessary to
restart an algorithm (besides psi, model, options).
This commit implements a method in Algorithm providing that data.

It breaks backwards compatibility for already for files saved by
previous Simulation classes, but this should be okay since the
Simulation class is not yet included in any official release.