fix another logging issue
The Config class raises warnings in its `__del__`, which can be delayed
until the program exits. The logging module can already have closed it's logfile at that stage.
However, it then *re-opens* the logfile to write that warning (provided
that captureWarnings is set). With file mode 'w', this *erases* previous log messages.

The best fix I came up with is to switch to file mode 'a', and to remove existing log files during initialization (of the simulation class).