INCOMPATIBLE FIX: measurement functions now need to accept `model` parameter
This is necessary to avoid special-cases inside measurment functions for
grouped sites, orthogonalExcitations and others. The model is guaranteed
to be compatible with the state.

This is a cherry-pick of a previous commit 2a5052ac2abc672b0f2e83d8f74fb3ca3be2229b
not yet merged in the main branch, but already existing other branches, in particular excitations_science and vumps.