make cons_Sz='parity' for SpinHalfFermionSite non-trivial again.
Reverts dc20d98997be91e6e4aad33c30f1a9044ffb7699, which claimed to "fix" it.
However, that "fix" yielded the same charges as cons_N='parity' conservation - and thus was trivial if the (parity of the) particle number is conserved.
The revert makes the charges non-trivial again, which has the potential to speed up simulations with cons_N='N', cons_Sz='parity'.
Seems like a somewhat unlikely combination for "physical" hamiltonians, though, so hopefully not too bad that this was there...?