fix memory leak when repeatedly calling MPS_common.init_env()
init_env() called self.cache.create_subcache('env') even when such a
cache was already created. For the trivial cache, this leads to a memory
leak, since the DictCache.long_term_storage kept a copy of the subcache
container.
For non-trivial caches, e.g. with a PickleStorage, this would actually
have raised an error....