provide proper measurement wrappers psi_method and simulation_method
The wrappers provide a simple way to measure not only global functions
but also methods of psi and the simulation class.

Here, the fix of 2a5052ac2abc672b0f2e83d8f74fb3ca3be2229b
is not yet included, so the next commit is a cherry-pick merging only that.