Make Y0 projection more robust
Have assumed that Y0._qdata[:, 0] is increasing (i.e. sorted).
Since the charges are sorted in the legs, Y0._qdata[:, -1] is indeed sorted.
For qmod==1 we then have _qdata[:, 0] == _qdata[:, -1] + shift_from_qtotal, which is also sorted.
For qmod>1, however, this is no longer true.
Now we explicitly use the order to pick the right blocks.