Fix regrouping in Simulation class.
If in a Simulation run we set the parameters `group_to_NearestNeighborModel`
to True and `group_sites` > 2, the Simulation measures as default
`energy_MPO`. However, the hamiltonian during the simulation is now
a NearestNeighbor Hamiltonian.
This commit removes `energy_MPO` as default measurement and adds
an additional method `get_extra_default_measurements` to the Model class.
Subclasses then define a list of additional (model dependent) default
measurements. The method gets called in the `_connect_measurements`
method of the simulation class.
In measurement.py, `m_bond_energies` uses then simulation.model and
simulation.psi