Include teste for BZ in lattice.py, turn Lattice.basis into property
Added tests to ensure correctness of the BZ calculation (with hardcoded solutions for the corners
of the Brillouin Zones of the default tenpy lattices).
Turn attribute basis in Lattice into a property, giving it an appropriate setter method.
This ensures that when the basis is changed manually, that the reciprocal lattice vectors and the BZ can still be calculated.
Furthermore introduce testing of BZ class and reciprocal_basis_vectors.