Move logic to generate a "tree" of possible charges from PurificationMPS into MPS class
In the classmethod from_infiniteT_canonical of PurificationMPS the first step corresponds
to finding a tree of possible charges at each site which corresponds to the desired charge sector.
This logic is moved into the MPS class as it will be used to project a product state onto a
desired charge sector. The classmethod from_infiniteT_canonical is adapted.