Add methods to project a product state onto a given charge sector for MPS
This is similar to the method from_infiniteT_canonical of the PurificationMPS class.
In the canonical ensemble each local basis state has the same probability in a charge sector.
For the new classmethod project_onto_charge_sector it is possible to pass a list of coefficients
for each local basis state for each site in the desired charge sector. This effectively projects
a product state onto a desired charge sector.