Incorporate changes after code-review.
For Simulations include a function `_display_errors_during_run` that displays errors
that occured during the run (i.e. in a post-processing step) which are stored in the
attribute errors_during_run.
Furthermore include the `default_post_processing` list as class-attribute in Simulation.
Overwrite `init_measurements` instead of `_connect_measurements` in class `TimeDependentCorrelation`