OrthogonalExcitation.switch_charge_sector for finite systems shouldn't use trivial bond
Warnings during tests enabled in #379 caused the test_simulation_exc.py
to fail due to a division-by-zero warning.
Indeed, in switch_charge_sector, we've been using the left-most bond,
which is trivial for finite systems, and did a zero-site Lanczos on it
to switch sector. But this can only work for non-trivial bonds, i.e. infinite/segment systems.
For finite systems, we have to do this on another bond - which this commit implements.